% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQSpline.R
\name{buildQSpline}
\alias{buildQSpline}
\title{Build a monotone quantile spline Q(u) for ln(Sa)}
\usage{
buildQSpline(SaTable)
}
\arguments{
\item{SaTable}{data.table with probability column \code{p} (0-1 or "mean")
and spectral-acceleration column \code{Sa}.}
}
\value{
A closure \code{Q(u)} returning \code{ln(Sa)} for any \code{u in[0,1]}.
}
\description{
Build a monotone quantile spline Q(u) for ln(Sa)
}
