% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUHS.R
\name{checkUHS}
\alias{checkUHS}
\title{Check monotonicity and duplicates in UHSTable}
\usage{
checkUHS(UHSTable, epsMonot = 1e-06)
}
\arguments{
\item{UHSTable}{data.table with columns
\code{TR, Vs30, Vref, ID, Tn, p, Sa}.}

\item{epsMonot}{numeric tolerance (default 1e-6).}
}
\value{
invisibly \code{TRUE}; stops on error.
}
\description{
Ensures that, for every combination
\code{TR, Vs30, Vref, ID, Tn}, there is exactly
one row per probability \code{p} and that
\code{Sa(p)} is non‑decreasing.
}
