% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDn.R
\name{Dn_AM88}
\alias{Dn_AM88}
\title{Ambraseys & Menu (1988) rigid sliding-block model}
\usage{
Dn_AM88(PGA, ky)
}
\arguments{
\item{PGA}{Peak ground acceleration (g).}

\item{ky}{Yield acceleration (g).}
}
\value{
data.table(muLnD, sdLnD, ID = "AM88")
}
\description{
Ambraseys & Menu (1988) rigid sliding-block model
}
