% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Dn_BM19}
\alias{Dn_BM19}
\title{Bray & Macedo (2019) – shallow-crustal events}
\usage{
Dn_BM19(ky, Ts, Sa, PGA, PGV, Mw = 6.5)
}
\arguments{
\item{ky}{Yield acceleration (g).}

\item{Ts}{Fundamental period (s).}

\item{Sa}{Spectral acceleration at 1.3 * Ts (g).}

\item{PGA}{Peak ground acceleration (g).}

\item{PGV}{Peak ground velocity (cm/s).}

\item{Mw}{Moment magnitude (default 6.5).}
}
\value{
data.table(muLnD, sdLnD, ID = "BM19")
}
\description{
Bray & Macedo (2019) – shallow-crustal events
}
