% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dn_models.R
\name{Dn_BT07}
\alias{Dn_BT07}
\title{Bray & Travasarou (2007) flexible sliding-block model}
\usage{
Dn_BT07(ky, Sa, Ts, Mw = 6.5)
}
\arguments{
\item{ky}{Yield acceleration (g).}

\item{Sa}{Spectral acceleration at 1.5 * Ts (g).}

\item{Ts}{Fundamental period (s).}

\item{Mw}{Moment magnitude.}
}
\value{
data.table(muLnD, sdLnD, ID = "BT07")
}
\description{
Bray & Travasarou (2007) flexible sliding-block model
}
