% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dn_models.R
\name{Dn_JB07}
\alias{Dn_JB07}
\title{Jibson (2007) empirical model}
\usage{
Dn_JB07(PGA, ky, AI = NULL)
}
\arguments{
\item{PGA}{Peak ground acceleration (g).}

\item{ky}{Yield acceleration (g).}

\item{AI}{Arias Intensity (m/s); if NULL, back-calculated from PGA.}
}
\value{
data.table(muLnD, sdLnD, ID = "JB07")
}
\description{
Jibson (2007) empirical model
}
