% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Dn_SR08}
\alias{Dn_SR08}
\title{Saygili & Rathje (2008) model}
\usage{
Dn_SR08(PGA, ky, AI = NULL)
}
\arguments{
\item{PGA}{Peak ground acceleration (g).}

\item{ky}{Yield acceleration (g).}

\item{AI}{Arias Intensity (m/s); if NULL, back-calculated from PGA.}
}
\value{
data.table(muLnD, sdLnD, ID = "SR08")
}
\description{
Saygili & Rathje (2008) model
}
