% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSaF.R
\name{F_ST17}
\alias{F_ST17}
\title{Non-linear site amplification factor \strong{F\if{html}{\out{<sub>}}ST17\if{html}{\out{</sub>}}}}
\usage{
F_ST17(PGA, Tn, vs30, vref = 760, Vl = 200, Vu = 2000)
}
\arguments{
\item{PGA}{numeric \strong{vector} – peak ground acceleration (in g).}

\item{Tn}{numeric \strong{scalar} – oscillator period (s).}

\item{vs30}{numeric \strong{scalar} – time-averaged shear-wave velocity to 30 m (m/s).}

\item{vref}{numeric \strong{scalar} – reference velocity (760 m/s or 3000 m/s), default 760.}

\item{Vl}{numeric scalar – lower Vs30 bound for NL sigma interpolation (default 200).}

\item{Vu}{numeric scalar – upper Vs30 bound for NL sigma interpolation (default 2000).}
}
\value{
A \link[data.table]{data.table} with columns:
\describe{
\item{\code{muLnF}}{natural-log mean of \eqn{F}}
\item{\code{sdLnF}}{natural-log standard deviation of \eqn{F}
(set to \code{NA} when \code{vs30 == vref})}
\item{\code{ID}}{character string \code{"ST17"}}}
}
\description{
Implements the period-dependent, Vs30-dependent model of Stewart & Seyhan (2017)
for the natural-log mean and standard deviation of the amplification factor
\eqn{F(T_n,\mathrm{PGA},V_{s30})}.
}
\details{
\emph{If \code{vs30 == vref}, the site factor equals 1, so the function returns
\eqn{\mu_{\ln F}=0} and \code{sdLnF = NA} (to signal deterministic behaviour).}
}
\section{References}{

Stewart, J.P. & Seyhan, E. (2017) “Semi-empirical nonlinear site
amplification model for global application.” \emph{Earthquake Spectra} \strong{33}(1).
}

