% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDn.R
\name{fitDn}
\alias{fitDn}
\title{Probabilistic Newmark displacements (Monte Carlo + epistemic mixing)}
\usage{
fitDn(uhs, ky, Ts, Mw = 6.5, NS = 30L, Rrup = 100)
}
\arguments{
\item{uhs}{data.table ― columns Tn, p, Sa  (must include p == "mean")}

\item{ky}{numeric scalar, yield acceleration (g)}

\item{Ts}{numeric scalar, fundamental period of sliding mass (s)}

\item{Mw}{numeric scalar, moment magnitude (default 6.5)}

\item{NS}{integer, number of Monte‑Carlo samples}

\item{Rrup}{numeric scalar, rupture distance for rhoBJ()}
}
\value{
data.table with columns p, Dn  (displacement in cm)
}
\description{
Probabilistic Newmark displacements (Monte Carlo + epistemic mixing)
}
