% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDn.R
\name{getDn}
\alias{getDn}
\title{Internal helper: draw n samples from a displacement model}
\usage{
getDn(.fun, ..., n = 1)
}
\description{
Internal helper: draw n samples from a displacement model
}
\keyword{internal}
