% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateSaTable.R
\name{interpolateSaTable}
\alias{interpolateSaTable}
\title{Interpolate Sa(p) at arbitrary period using log-log interpolation}
\usage{
interpolateSaTable(uhs, Tn)
}
\arguments{
\item{uhs}{data.table with columns Tn, p, Sa.}

\item{Tn}{numeric scalar — target period for interpolation.}
}
\value{
data.table with interpolated Tn, p, Sa
}
\description{
Interpolate Sa(p) at arbitrary period using log-log interpolation
}
