% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSaF.R
\name{SaF_SS17}
\alias{SaF_SS17}
\title{Site–factor model SaF_ST17}
\usage{
SaF_SS17(Sa, PGA, Tn, vs30, vref = 760)
}
\arguments{
\item{Sa}{numeric vector – sampled spectral acceleration at Tn (g).}

\item{PGA}{numeric vector – sampled peak ground acceleration (g).}

\item{Tn}{numeric scalar – oscillator period (s).}

\item{vs30}{numeric scalar – time-averaged Vs30 (m/s).}

\item{vref}{numeric scalar – reference Vs (default 760 m/s).}
}
\value{
data.table(muLnSaF, sdLnSaF, ID = "ST17")
}
\description{
Computes median (muLnSaF) and log-standard deviation (sdLnSaF) of
\deqn{SaF = Sa(T_n)\;×\;F_{ST17}(T_n,\mathrm{PGA},V_{s30})}.
}
