% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSa.R
\name{sampleSa}
\alias{sampleSa}
\title{Sample spectral acceleration consistent with UHS quantiles}
\usage{
sampleSa(UHS, Td, n)
}
\arguments{
\item{UHS}{data.table with Tn, Sa, p (prob labels + "mean").}

\item{Td}{numeric scalar – target period (s).}

\item{n}{integer – number of samples.}
}
\value{
list(SaTable, muLnSa, sdLnSa, Sa)
}
\description{
Sample spectral acceleration consistent with UHS quantiles
}
