% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSaQ.R
\name{sdSaQ}
\alias{sdSaQ}
\title{Robust period-independent ln-σ for an Sa spectrum}
\usage{
sdSaQ(UHS)
}
\arguments{
\item{UHS}{data.table with Tn, Sa, p  (must include row p == "mean").}
}
\value{
numeric scalar  —  σ*  (ln units).
}
\description{
Robust period-independent ln-σ for an Sa spectrum
}
\keyword{internal}
