% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDn.R
\name{Dn_YG91}
\alias{Dn_YG91}
\title{Yegian et al. (1991) rigid sliding-block model}
\usage{
Dn_YG91(PGA, ky)
}
\arguments{
\item{PGA}{Peak ground acceleration (g).}

\item{ky}{Yield acceleration (g).}
}
\value{
data.table(muLnD, sdLnD, ID = "YG91")
}
\description{
Yegian et al. (1991) rigid sliding-block model
}
