% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSaF.R
\name{SaF_ST17}
\alias{SaF_ST17}
\title{Site–factor model SaF_ST17  (canonical)}
\usage{
SaF_ST17(Sa, pga, Tn, vs30, vref = 760)
}
\arguments{
\item{Sa}{numeric – sampled spectral acceleration at Tn (g)}

\item{pga}{numeric – sampled peak ground acceleration (g)}

\item{Tn}{numeric scalar – oscillator period (s)}

\item{vs30}{numeric scalar – time-averaged Vs30 (m/s)}

\item{vref}{numeric scalar – reference Vs30 (default 760 m/s)}
}
\value{
data.table(muLnSaF, sdLnSaF, ID = "ST17")
}
\description{
Site–factor model SaF_ST17  (canonical)
}
